+++ vs2022/Platform/Elements/22_Titanium.h
@@ -12,7 +12,7 @@ namespace Dysnomia {
 			{
 			public:
 				static Quark<Dysnomium^, Scandium^, Calcium^>^ Despina;
-				BigInteger Ring;
+				Int64 Ring;
 
 				Titanium(Calcium^ Upsilon, Scandium^ Nu) : Quark(Nu, Upsilon, Upsilon->Mu.Key->N) {
 					if (Despina == nullptr) Despina = gcnew Quark<Dysnomium^, Scandium^, Calcium^>(L, N, R);
